refactor: adding errors to use cases
