+++ src/use-cases/errors/customer-already-exists-error.ts
@@ -1,0 +1,5 @@
+export class CustomerAlreadyExistsError extends Error {
+    constructor() {
+      super('cpf already registered')
+    }
+  }
