+++ src/use-cases/errors/customer-not-found-error.ts
@@ -1,0 +1,5 @@
+export class CustomerNotFound extends Error {
+    constructor() {
+      super('customer not found')
+    }
+  }
