+++ src/use-cases/register-customers.ts
@@ -1,5 +1,6 @@
 import { customers } from "@prisma/client"
 import { CustomersRepository } from "../repositories/customers-repository"
+import { CustomerAlreadyExistsError } from "./errors/customer-already-exists-error"
 
 interface RegisterCustormesRequest{
   cpf : string
@@ -23,6 +24,12 @@ export class RegisterCustormesUseCase{
         name, 
         numberPhone
     }: RegisterCustormesRequest): Promise<RegisterCustormesResponse> {
+
+        const customerWithExistingCpf = await this.customersRepository.findyByCpf(cpf)
+
+        if(customerWithExistingCpf){
+            throw new CustomerAlreadyExistsError()
+        }
     
         const customers = await this.customersRepository.create({
             cpf,
